refactor: minor tweaks
